featL Add interiorDest to XcmCreator
