+++ src/createXcmTypes/xcm/v3.ts
@@ -1,11 +1,14 @@
 import { AnyJson } from '@polkadot/types-codec/types';
 import { isEthereumAddress } from '@polkadot/util-crypto';
 
+import { BaseError, BaseErrorsEnum } from '../../errors/BaseError.js';
 import { RemoteReserve } from '../../types.js';
 import { sanitizeKeys } from '../../util/sanitizeKeys.js';
 import {
 	FungibleAssetType,
 	FungibleMultiAsset,
+	InteriorKey,
+	InteriorValue,
 	UnionXcAssetsMultiAssets,
 	UnionXcAssetsMultiLocation,
 	UnionXcmMultiLocation,
@@ -138,6 +141,28 @@ export const V3: XcmCreator = {
 			V3: {
 				parents,
 				interior: { Here: null },
+			},
+		};
+	},
+
+	interiorDest({ destId, parents }: { destId: string; parents: number }): XcmDestBeneficiary {
+		const multiLocation = parseLocationStrToLocation(destId);
+
+		let interior: InteriorKey | undefined = undefined;
+		if (multiLocation && multiLocation.interior.X1) {
+			interior = { X1: multiLocation.interior.X1 as InteriorValue };
+		} else {
+			interior = { X2: multiLocation.interior.X2 as InteriorValue };
+		}
+
+		if (!interior) {
+			throw new BaseError('Unable to create XCM Destination location', BaseErrorsEnum.InternalError);
+		}
+
+		return {
+			V3: {
+				parents,
+				interior,
 			},
 		};
 	},
